Salvador

+++ Application/Services/PolicyServices.cs
@@ -188,46 +188,60 @@ namespace Application.Services
                 policy.ExpirationDate = request.ExpirationDate ?? policy.ExpirationDate;
                 policy.NameCompany = request.NameCompany ?? policy.NameCompany;
 
-                var policyIma = await _unitOfWork.PhotosOfPolicyRepo.Get(filter: policy => policy.PolicyId == request.PolicyId);    
-                foreach ( var photo in policyIma)
+                if(request.Images.Count > 0)
                 {
-                    await _blobStorageService.DeleteFileFromBlobAsync(_azureBlobContainers.Value.PolicyImages, photo.FilePath);
-                    await _unitOfWork.PhotosOfPolicyRepo.Delete(photo.Id);
-                }
-
-                var Ima = new List<PhotosOfPolicy>();
-                foreach (var image in request.Images)
-                {
-                    //Validar Imagenes y Guardar las imagenes en el blobstorage
-                    if (image.ContentType.Contains("pdf"))
+                    var policyIma = await _unitOfWork.PhotosOfPolicyRepo.Get(filter: policy => policy.PolicyId == request.PolicyId);
+                    foreach (var photo in policyIma)
                     {
-                        //Manipular el nombre del archivo
-                        var uploadDate = DateTime.Now;
-                        Random rndm = new Random();
-                        string FileExtn = System.IO.Path.GetExtension(image.FileName);
-                        var filePath = $"FOTOS_POLIZA/{uploadDate.Day}{uploadDate.Month}{uploadDate.Year}_{uploadDate.Hour}{uploadDate.Minute}{rndm.Next(1, 1000)}{FileExtn}";
-                        var uploadedUrl = await _blobStorageService.UploadFileToBlobAsync(image, _azureBlobContainers.Value.PolicyImages, filePath);
+                        await _blobStorageService.DeleteFileFromBlobAsync(_azureBlobContainers.Value.PolicyImages, photo.FilePath);
+                        await _unitOfWork.PhotosOfPolicyRepo.Delete(photo.Id);
+                    }
 
-                        //agregar la imagen a la bd
-                        var newImage = new PhotosOfPolicy()
+                    var Ima = new List<PhotosOfPolicy>();
+                    foreach (var image in request.Images)
+                    {
+                        //Validar Imagenes y Guardar las imagenes en el blobstorage
+                        if (image.ContentType.Contains("pdf"))
                         {
-                            FilePath = filePath,
-                            FileURL = await _blobStorageService.GetFileUrl(_azureBlobContainers.Value.PolicyImages, filePath),
-                            Policy = policy
-                        };
+                            //Manipular el nombre del archivo
+                            var uploadDate = DateTime.Now;
+                            Random rndm = new Random();
+                            string FileExtn = System.IO.Path.GetExtension(image.FileName);
+                            var filePath = $"FOTOS_POLIZA/{uploadDate.Day}{uploadDate.Month}{uploadDate.Year}_{uploadDate.Hour}{uploadDate.Minute}{rndm.Next(1, 1000)}{FileExtn}";
+                            var uploadedUrl = await _blobStorageService.UploadFileToBlobAsync(image, _azureBlobContainers.Value.PolicyImages, filePath);
 
-                        await _unitOfWork.PhotosOfPolicyRepo.Add(newImage);
-                        Ima.Add(newImage);
-                    }
-                    else
-                    {
-                        response.success = false;
-                        response.AddError("Archivo de Imagen Invalido", "Uno o mas archivos no corresponden a un archivo de Imagen", 5);
+                            //agregar la imagen a la bd
+                            var newImage = new PhotosOfPolicy()
+                            {
+                                FilePath = filePath,
+                                FileURL = await _blobStorageService.GetFileUrl(_azureBlobContainers.Value.PolicyImages, filePath),
+                                Policy = policy
+                            };
 
-                        return response;
+                            await _unitOfWork.PhotosOfPolicyRepo.Add(newImage);
+                            Ima.Add(newImage);
+                        }
+                        else
+                        {
+                            response.success = false;
+                            response.AddError("Archivo de Imagen Invalido", "Uno o mas archivos no corresponden a un archivo de Imagen", 5);
+
+                            return response;
+                        }
+
                     }
 
                 }
+                else
+                {
+                    await _unitOfWork.SaveChangesAsync();
+                    response.success = true;
+                    var Dto = _mapper.Map<PolicyDto>(policy);
+                    response.Data = Dto;
+                    return response;
+
+                }
+      
 
                 await _unitOfWork.PolicyRepo.Update(policy);
                 await _unitOfWork.SaveChangesAsync();
